Fix concurrent update warning by extend submit event
Before this commit when submit is done by Portal form-view with module
installed, you will see in the log:

`Error: could not serialize access due to concurrent update`

This happened because the query to setup the value in backend
of field `zip_id` doesn't need validation so it's not managed
by `/my/account` route validations: it is directly updated
by an rpc which calls a different json route. This means that
two diffent transaction happens, which is fine as long as they
are kept split and asyncronus.

In this commit I defined an handler that will directly catch
the submit event: now the the query to set `zip_id` will be
executed asyncronously, after the main transaction is completed.