Fix singleton error on archived products
When a product is archived, the product template loses
reference to its variants (until it gets unarchived).

This will cause issues, expecially for pricelist-rules based
on supplierinfo prices: method `_select_seller()` is only
defined on `product.product` model. This means that if
the method to compute prices is invoked on a template, it will
search for the `product_variant_ids[0]` to invoke `product.product`
functions `_select_seller()`, but function will get an empty recordset
as `self`, and it will make GUI crash since the `_select_seller()`
function invokes `self.ensure_one()` method.

This commit will not really solve the issue, but it will just skip
the computation of prices for archived products, so the GUI will
not crash when you have an archived product on a "product template
based" pricelist-item.

Warnings are added:

- in the syslog
- on pricelist item form-view

to inform users that the price wil not be computed until
the product is archived.

- a specific `odoo.exceptions.UserError` will pop when
user press button to set % discount on an archived product.

* A technical computed Boolean field named `unable_to_retrieve_variant`
has been added to be able to show an "alert div" on the form-view.

* An hook `_reset_computed_vals()` has been added to make the default
values computation a bit more clean.