Avoid rounding inconsistencies between SO and Rule for rules based on… … "other pricelists"
To compute the price of the product in the context
of a rule (instead of saleorder) I use custom method
`_fetch_supplierinfo_price()` because I need to fetch
seller price before the rule is applied, plus
`_get_supplierinfo_pricelist_price()` wouldn't be
GUI responsive in the pricelist/pricelist rule context.

This method can also performa a currency conversion
if needed, but the `_convert()` method required an
explicit arg `round=False` to be aligned with the
rounding applied on sale order context.

In sale order context the rounding factor is defined by
`digits='Product Price'` for fields like `price_unit`,
or by decimal accuracy assigned to the currency for
monetary fields like `price_total`, `price_subtotal`.

The precision digit assigned by default from these parameters
is 2, BUT for rules based on "other pricelists", there is
a step where this rounding is not applied: when the "base"
pricelist price is computed by `_compute_price_rule()`.

In this step the "base" pricelist price for the rule is
recomputed without rounding, and we need to do the same
in our custom method to align it with saleorder, otherwise
the computed price would show a slightly different price
than `price_unit` in a sale order, due to the rounding factor.

Also, I realized that `_convert()` method was called improperly
(called also when pricelist or company currency is the same as
seller currency): now the rounding is not applied on convert,
and convert is only called when conversion is required.

The _convert() called in `_get_supplierinfo_pricelist_price()`
still does not set `round=False`, so if a rule:

- is based on a pricelist
- the base pricelist is based on supplierinfo prices
- the seller currency is different than pricelist (or company) currency

there might still be inconsistency between computed price in the context
of a pricelist, and the computed price in the context of a rule.