Add zip_id to Road Map
Detected an issue with _onSubmit method: it was referenced as event
throught an xml selector, but it caused the event not to be triggered.

When I actually managed it to trigger the `zip_id` query on submit event
I realized that asyncronous query has some issues in this context.

At the moment I removed the `zip_id` setup from front-end side and
commented the respective section of code with some info.